backend/requirements.txt
openai>=1.12.0
langchain>=0.1.9
langchain-core>=0.1.27
langchain-community>=0.0.24
langchain-openai>=0.0.8
langgraph>=0.0.26
python-dotenv>=1.0.1
tavily-python>=0.3.1
langgraph-cli[inmem]>=0.0.5
langgraph-sdk>=0.0.5